requirements.txt
django-organizations>=0.9.0,<0.10.0
django-extensions>=1.7.5
django-allauth>=0.30.0,<0.31.0
